src/requirements.txt
Pillow
numpy
opencv-python
scikit-build